회고 상세내용 컴포넌트 분리하라
